// TODO not use REFLECTION in OSGI
In OSGi you will want to stay away from reflection, for reasons outlined in a lot of other places.

So, your situation is that bundle A needs some instance of a class which resides in bundle B. For A to make sense of this instance, I will assume it has some interface that it will use to talk to the instance. Let's make this a little more concrete.

/Bundle A
  /ThingyInterface.class

/Bundle B
  /ThingyImplementation.class (implements ThingyInterface.class)
This is a regular pattern: one bundle provides an interface, the other provides the implemenatation. There are now two possible situations,

A needs exactly one copy of the implementation. In that case, register the Thingy as a service.
A needs several instances of the implementation. In that case, introduce a ThingyFactory in A, and create an implementation of that factory in B, which you then register as a service.
In either case, you let B do the actual instantiation, you have no dependency from A to B, and B doesn't need reflection to instantiate the objects.

In short, services are your friend.